requirements.txt
langgraph
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-groq
python-dotenv
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4